pyproject.toml
[project]
name = "kcn2"
version = "0.1.0"
description = ""
authors = [{ name = "''" }]
readme = "README.md"
requires-python = "==3.13.1"
dependencies = [
    "result (==0.17.0)",
    "orjson (==3.10.15)",
    "aiohttp[speedups] (==3.11.11)",
    "loguru (==0.7.3)",
    "websockets (==14.2)",
]

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = ["N815"]
lint.fixable = ["I", "COM"]
lint.mccabe.max-complexity = 3
lint.pydocstyle.convention = "google"

target-version = "py313"

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.4"
mypy = "^1.15.0"
pytest-cov = "^6.0.0"
vulture = "^2.14"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
